Plot experience study results

Plot experience study results

Usage

# S3 method for exp_df
autoplot(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  mapping,
  scales = "fixed",
  geoms = c("lines", "bars"),
  y_labels = scales::label_percent(accuracy = 0.1)
)

# S3 method for trx_df
autoplot(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  mapping,
  scales = "fixed",
  geoms = c("lines", "bars"),
  y_labels = scales::label_percent(accuracy = 0.1)
)

Arguments

object

An object of class exp_df created by the function exp_stats() or an object of class trx_df created by the function trx_stats().

...

Faceting variables passed to ggplot2::facet_wrap().

x

An unquoted column name in object or expression to use as the x variable.

y

An unquoted column name in object or expression to use as the y variable. If unspecified, y will default to the observed termination rate (q_obs) for exp_df objects and the observed utilization rate (trx_util) for trx_df objects.

color

An unquoted column name in object or expression to use as the color and fill variables.

mapping

Aesthetic mapping passed to ggplot2::ggplot(). NOTE: If mapping is supplied, the x, y, and color arguments will be ignored.

scales

The scales argument passed to ggplot2::facet_wrap().

geoms

Type of geometry. If "points" is passed, the plot will display lines and points. If "bars", the plot will display bars.

y_labels

Label function passed to ggplot2::scale_y_continuous().

Value

a ggplot object

Details

If no aesthetic map is supplied, the plot will use the first grouping variable in object on the x axis and q_obs on the y axis. In addition, the second grouping variable in object will be used for color and fill.

If no faceting variables are supplied, the plot will use grouping variables 3 and up as facets. These variables are passed into ggplot2::facet_wrap(). Specific to trx_df objects, transaction type (trx_type) will also be added as a faceting variable.