Summarize experience study records

Create a summary data frame of experience for a given target status.

Usage

exp_stats(
  .data,
  target_status = attr(.data, "target_status"),
  expected,
  col_exposure = "exposure",
  col_status = "status",
  wt = NULL,
  credibility = FALSE,
  cred_p = 0.95,
  cred_r = 0.05
)

# S3 method for exp_df
summary(object, ...)

Arguments

.data

a data frame with exposure-level records, ideally of type exposed_df

target_status

a character vector of target status values

expected

a character vector containing column names in .data with expected values

col_exposure

name of the column in .data containing exposures

col_status

name of the column in .data containing the policy status

wt

Optional. Length 1 character vector. Name of the column in .data containing weights to use in the calculation of claims, exposures, and partial credibility.

credibility

whether the output should include partial credibility weights and credibility-weighted decrement rates.

cred_p

confidence level under the Limited Fluctuation credibility method

cred_r

error tolerance under the Limited Fluctuation credibility method

object

an exp_df object

...

groups to retain after summary() is called

Value

A tibble with class exp_df, tbl_df, tbl, and data.frame. The results include columns for any grouping variables, claims, exposures, and observed decrement rates (q_obs). If any values are passed to expected, additional columns will be added for expected decrements and actual-to-expected ratios. If credibility is set to TRUE, additional columns are added for partial credibility and credibility-weighted decrement rates (assuming values are passed to expected).

Details

If .data is grouped, the resulting data frame will contain one row per group.

If target_status isn't provided, exp_stats() will use the same target status from .data if it has the class exposed_df. Otherwise, .data is not an exposed_df object, all status values except the first level will be assumed. This will produce a warning message.

Expected values

The expected argument is optional. If provided, this argument must be a character vector with values corresponding to columns in .data containing expected experience. More than one expected basis can be provided.

Credibility

If credibility is set to TRUE, the output will contain a credibility column equal to the partial credibility estimate under the Limited Fluctuation credibility method (also known as Classical Credibility) assuming a binomial distribution of claims.

summary() Method

Applying summary() to a exp_df object will re-summarize the data while retaining any grouping variables passed to the "dots" (...).

References

Herzog, Thomas (1999). Introduction to Credibility Theory

Examples

toy_census |> expose("2020-12-31", target_status = "Surrender") |>
    exp_stats()
#> Experience study results
#> 
#>  Groups:  
#>  Target status: Surrender 
#>  Study range: 1900-01-01 to 2020-12-31 
#> 
#> # A tibble: 1 × 4
#>   n_claims claims exposure q_obs
#> *    <int>  <int>    <dbl> <dbl>
#> 1        0      0     31.4     0

exp_res <- census_dat |>
           expose("2019-12-31", target_status = "Surrender") |>
           dplyr::group_by(pol_yr, inc_guar) |>
           exp_stats()

exp_res
#> Experience study results
#> 
#>  Groups: pol_yr, inc_guar 
#>  Target status: Surrender 
#>  Study range: 1900-01-01 to 2019-12-31 
#> 
#> # A tibble: 30 × 6
#>    pol_yr inc_guar n_claims claims exposure   q_obs
#>  *  <int> <lgl>       <int>  <int>    <dbl>   <dbl>
#>  1      1 FALSE          42     42    7719. 0.00544
#>  2      1 TRUE           38     38   11526. 0.00330
#>  3      2 FALSE          65     65    7117. 0.00913
#>  4      2 TRUE           61     61   10605. 0.00575
#>  5      3 FALSE          67     67    6476. 0.0103 
#>  6      3 TRUE           60     60    9626. 0.00623
#>  7      4 FALSE         103    103    5823. 0.0177 
#>  8      4 TRUE           58     58    8697. 0.00667
#>  9      5 FALSE          94     94    5147. 0.0183 
#> 10      5 TRUE           72     72    7779. 0.00926
#> # … with 20 more rows
summary(exp_res)
#> Experience study results
#> 
#>  Groups:  
#>  Target status: Surrender 
#>  Study range: 1900-01-01 to 2019-12-31 
#> 
#> # A tibble: 1 × 4
#>   n_claims claims exposure  q_obs
#> *    <int>  <int>    <dbl>  <dbl>
#> 1     2846   2846  132669. 0.0215
summary(exp_res, inc_guar)
#> Experience study results
#> 
#>  Groups: inc_guar 
#>  Target status: Surrender 
#>  Study range: 1900-01-01 to 2019-12-31 
#> 
#> # A tibble: 2 × 5
#>   inc_guar n_claims claims exposure  q_obs
#> * <lgl>       <int>  <int>    <dbl>  <dbl>
#> 1 FALSE        1615   1615   52338. 0.0309
#> 2 TRUE         1231   1231   80331. 0.0153